requirements.txt
re
csv
json
faiss
requests
numpy
pandas
defaultdict
SentenceTransformer
CrossEncoder